+++ EmrCloudApi/Requests/ExportPDF/YakutaiRequest.cs
@@ -4,5 +4,5 @@ public class YakutaiRequest
 {
     public long PtId { get; set; }
     public int SinDate { get; set; }
-    public int RaiinNo { get; set; }
+    public long RaiinNo { get; set; }
 }
