+++ EmrCloudApi/Requests/SwapHoken/CalculationSwapHokenRequest.cs
@@ -2,7 +2,7 @@
 {
     public class CalculationSwapHokenRequest
     {
-        public CalculationSwapHokenRequest(List<int> seikyuYms, int ptId, bool isReCalculation, bool isReceCalculation, bool isReceCheckError)
+        public CalculationSwapHokenRequest(List<int> seikyuYms, long ptId, bool isReCalculation, bool isReceCalculation, bool isReceCheckError)
         {
             SeikyuYms = seikyuYms;
             PtId = ptId;
@@ -13,7 +13,7 @@
 
         public List<int> SeikyuYms { get; private set; }
 
-        public int PtId { get; private set; }
+        public long PtId { get; private set; }
 
         #region caculate module client
         public bool IsReCalculation { get; private set; }
